requirements.txt
# Web framework
fastapi==0.135.1
uvicorn[standard]==0.34.0

# HTTP client
httpx==0.28.1

# Database
chromadb==0.6.3

# Math
sympy==1.13.3

# Utilities
python-multipart==0.0.22
pydantic>=2.11.0,<3.0.0

# Channels
discord.py==2.4.0
python-telegram-bot==21.10

# Calendar
ics==0.7.2
tatsu==5.8.3

# Browser automation
playwright==1.49.1

# Desktop automation (optional — only used when ENABLE_DESKTOP_AUTOMATION=true)
pyautogui==0.9.54

# Voice (optional — only used when ENABLE_VOICE=true)
openai-whisper==20250625

# MCP (Model Context Protocol)
mcp>=1.23.0

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
